Update ALL_MULTIPLE_KEYBOARD_ASSIGNMENTS.ahk
FUNCTION FOR DIRECTLY APPLYING A PRESET EFFECT TO A CLIP! 

With these changes, I was able to make it work in my Premiere 2025. How I managed it was pure luck, because I understand almost nothing about scripting. It seems that after the Send %item%, it needs to return focus to the "Effects" panel for the next steps to work properly.
It still worked even after I moved the window around and switched workspaces!
I don't know if you're still using it, but I'm really glad I found this script — it's going to help me a lot.